Fix SPI so that SPI4 Clock will not be 2 times requested
The clocks assigned to the SPI blocks spi4 to spi6  were using the APB1 prescaler (42 mHZ) constant and should have been the APB2 prescaler (84 mHZ) constant. Thus the actual SPI clock on the bus was 2 times what is should have been (After adjustment for requested to actual)

IE: Request for 11 mHz maps to an Actual 10.5 mHz but would have produced a clock of 21 mHz on the bus.

This change fixes that.